requirements.txt
absl-py==0.12.0
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.4
astroid==2.4.2
astunparse==1.6.3
async-timeout==3.0.1
attrs==20.3.0
autobahn==21.3.1
Automat==20.2.0
autopep8==1.5.4
beautifulsoup4==4.9.1
cachetools==4.2.2
certifi==2020.6.20
cffi==1.14.5
channels==2.4.0
channels-redis==3.3.0
chardet==3.0.4
charset-normalizer==2.0.3
click==7.1.2
colorama==0.4.3
colorgram.py==1.2.0
comtypes==1.1.7
constantly==15.1.0
cryptography==3.4.7
cssselect==1.1.0
cx-Freeze==6.5.3
cycler==0.10.0
Cython==0.29.21
daphne==2.5.0
distlib==0.3.1
Django==3.2.4
django-3-jet==1.0.8
django-extensions==3.1.3
django-jet==1.0.8
django-redis==5.0.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.8.0
docutils==0.16
filelock==3.0.12
flatbuffers==1.12
fuzzywuzzy==0.18.0
gast==0.3.3
gensim==4.0.1
google-auth==1.30.0
google-auth-oauthlib==0.4.4
google-pasta==0.2.0
graphviz==0.16
grpcio==1.32.0
gunicorn==20.1.0
h2==3.2.0
h5py==2.10.0
hiredis==2.0.0
hpack==3.0.0
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
importlib-metadata==3.10.0
imutils==0.5.3
incremental==21.3.0
isort==5.5.1
itemadapter==0.2.0
itemloaders==1.0.4
jet==1.0
jmespath==0.10.0
joblib==0.16.0
Keras==2.4.3
Keras-Preprocessing==1.1.2
Kivy==2.0.0
kivy-deps.angle==0.3.0
kivy-deps.glew==0.3.0
kivy-deps.sdl2==0.3.1
Kivy-Garden==0.1.4
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
lxml==4.6.3
Markdown==3.3.4
matplotlib==3.3.2
mccabe==0.6.1
MouseInfo==0.1.3
msgpack==1.0.2
nltk==3.5
numpy==1.19.2
oauthlib==3.1.0
opencv-contrib-python==4.4.0.44
opencv-python==4.4.0.44
opt-einsum==3.3.0
pandas==1.2.4
parsel==1.6.0
Pillow==8.2.0
prettytable==2.1.0
priority==1.3.0
Protego==0.1.16
protobuf==3.16.0
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
PyAudio @ file:///C:/Users/anura/Downloads/PyAudio-0.2.11-cp38-cp38-win_amd64.whl
PyAutoGUI==0.9.52
pycodestyle==2.6.0
pycparser==2.20
PyDispatcher==2.0.5
pydotplus==2.0.2
PyGetWindow==0.0.9
Pygments==2.7.3
PyJWT==2.1.0
pylint==2.6.0
PyMsgBox==1.0.8
pyOpenSSL==20.0.1
pyparsing==2.4.7
pyperclip==1.8.0
pypiwin32==223
PyRect==0.1.4
PyScreeze==0.1.26
python-dateutil==2.8.1
python-decouple==3.4
python-Levenshtein-wheels==0.13.2
pyttsx3==2.90
PyTweening==1.0.3
pytz==2021.1
pywin32==228
PyYAML==5.4.1
queuelib==1.6.1
razorpay==1.2.0
redis==3.5.3
regex==2020.7.14
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
scikit-learn==0.23.2
scipy==1.5.2
Scrapy==2.5.0
selenium==3.141.0
service-identity==18.1.0
six==1.15.0
smart-open==5.1.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
spotipy==2.18.0
sqlparse==0.4.1
tensorboard==2.5.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.4.1
tensorflow-estimator==2.4.0
termcolor==1.1.0
threadpoolctl==2.1.0
toml==0.10.1
torch==1.8.1+cu111
torchaudio==0.8.1
torchvision==0.9.1+cu111
tqdm==4.49.0
twilio==6.61.0
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
typing-extensions==3.7.4.3
urllib3==1.26.6
virtualenv==20.2.2
virtualenvwrapper-win==1.2.6
w3lib==1.22.0
wcwidth==0.2.5
Werkzeug==1.0.1
wikipedia==1.4.0
wordcloud==1.8.1
wrapt==1.12.1
zipp==3.4.1
zope.interface==5.4.0
